* break Video.tsx into smaller files
